Finished tasks and dependencies with colour code
The tasks are done with hopefully the time allocated properly. There is
a few days time between the end of implementation and submitting the
project so we can change the actual time taken as we see fit. I haven't
finished the resources, as I am not 100% on who is doing what.